handle basic map parsing cases

+++ src/utils/libft_tools_0.c
@@ -6,15 +6,15 @@
 /*   By: kben-tou <kben-tou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/19 15:29:48 by kben-tou          #+#    #+#             */
-/*   Updated: 2025/05/20 16:33:59 by kben-tou         ###   ########.fr       */
+/*   Updated: 2025/05/22 10:37:55 by kben-tou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/cub3d.h"
+#include "../../inc/cub3d.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!s)
